refactor: move ActiveSequences from lib/llm to lib/kv-router
Move the core per-worker sequence manager (ActiveSequences) into
dynamo-kv-router where it belongs alongside the other KV routing
primitives. ActiveSequencesMultiWorker stays in lib/llm since it
depends on dynamo_runtime event plane and Prometheus metrics.

Also fix pre-existing cfg gating bug in kv-router/indexer.rs where
DynamoError was imported unconditionally but dynamo-runtime is
feature-gated behind "metrics".
